Display a loading component (Issue #2)
Add a configuration object as second parameter.
It takes two properties:
Loader: A component
prop: A prop name that should be a boolean which determine when to display the Loader component
