+++ package.json
@@ -4,6 +4,9 @@
   "description": "High order component to call a load function if present.",
   "main": "index.js",
   "dependencies": {},
+  "peerDependencies": {
+    "react": "^15.3.0"
+  },
   "devDependencies": {
     "babel-cli": "^6.11.4",
     "babel-core": "^6.13.2",
@@ -19,7 +22,6 @@
     "eslint-plugin-import": "^1.12.0",
     "eslint-plugin-jsx-a11y": "^2.0.1",
     "eslint-plugin-react": "^6.0.0",
-    "react": "^15.3.0",
     "webpack": "^1.13.1"
   },
   "scripts": {
